requirements.txt
Scrapy == 2.8.0
python-scrapyd-api == 2.1.2
scrapyd == 1.4.1
scrapy-user-agents == 0.1.1

